Build the sources without compiler warnings
The functions defined in headers were `static`, so every translation unit that
included one without calling it warned that the definition was unused. They are
`inline` now, which is what a header wants anyway: one entity across the program
rather than a copy per TU. `BaseArgs` in the tests, which lives in an anonymous
namespace, gets `[[maybe_unused]]` instead.

weights.c holds two definitions and no code, and it was the only C file in the
build, so the C++-only flags on the warning line -- -Wnon-virtual-dtor,
-fpermissive, -fvisibility-inlines-hidden -- were handed to the C compiler on
every build and complained. It is weights.h now, with the array and its length
`constexpr`, and force_start.h includes it rather than declaring the two symbols
`extern` at the point of use. `size_wC20_5p_3p` being `std::size_t` also matches
the `size()` it is compared against.

printMat and printfloatMat take their sequences as `const unsigned char*`, which
is what the callers hold.

The energy divisions and the `%f` arguments in the force-start path convert a
float to double explicitly. Both already promoted -- the divisor is a double
literal, and a varargs float promotes -- so the value printed is unchanged and
-Wdouble-promotion has nothing left to say.

The index and length variables that are compared against a size take an unsigned
type, and RIs_linSpace stops narrowing the two sequence lengths on the way in.
The sweep and the window fill carry `std::size_t` for the target length to match.

Verified against the previous commit by building both and comparing output over
359 command-line cases and 500 randomized force-start alignments: identical
stdout, stderr and exit status throughout, and identical sanitizer reports under
-fsanitize=address,undefined.

+++ RIsearch1/src/align/HitReporter.h
@@ -60,7 +60,8 @@ public:
             m_matrices.Ix(), m_matrices.Iy(), m_profile, w.qbeg - 1, m_best.get());
 
         const auto energy =
-            (m_hit.max + m_config.extension_penalty * m_hit.nucleotide_count() - m_reference) /
+            static_cast<double>(m_hit.max + m_config.extension_penalty * m_hit.nucleotide_count() -
+                                m_reference) /
             (-100.0);
 
         if (energy <= m_config.max_energy) {
